pyproject.toml
[tool.poetry]
name = "tweet_collector"
version = "0.1.0"
description = ""
authors = ["parisa-zahedi <p.zahedi@uu.nl>","roos-voorvaart <r.voorvaart@uu.nl>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.4"
requests = "^2.25.1"
requests-oauthlib = "^1.3.0"
argparse = "^1.4.0"
searchtweets-v2 = "^1.1.0"
python-dateutil = "^2.8.1"
PyYAML = "^5.4.1"
elasticsearch = "^7.14.0"
ipykernel = "^6.4.1"

[tool.poetry.scripts]
tweet_collector="tweet_collector.search_tweet:main"
tweet_collector_elastic="tweet_collector.load_elastic:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
